+++ changelog.md
@@ -1,3 +1,8 @@
+v1.0.2 (26-10-2022)
+What's new?
+- The embeds for ping, yt and rankall commands do now contain the bot name and logo.
+- Added a gametest module update notify feature. It sends an embed in the news channel whenever a gametest module gets updated.
+
 v1.0.1 (23-10-2022)
 What's new?
 - Added roleAll command for administrators.